# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A small in-memory CRUD API for managing a to-do list. Built for W2 · A1."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev server" }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "description": "Returns basic information about this API and its endpoints.",
        "responses": {
          "200": { "description": "API info", "content": { "application/json": {} } }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns { status: 'ok' } if the server is alive.",
        "responses": {
          "200": { "description": "Server is healthy", "content": { "application/json": {} } }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "description": "Returns all tasks. Supports optional filtering, search, and pagination via query params.",
        "parameters": [
          { "name": "done", "in": "query", "schema": { "type": "boolean" }, "description": "Filter by completion status" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Filter by title substring (case-insensitive)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Max number of tasks to return" },
          { "name": "offset", "in": "query", "schema": { "type": "integer" }, "description": "Number of tasks to skip" }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } }
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Creates a new task. title is required and cannot be empty.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" },
              "example": { "title": "Buy milk" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "400": {
            "description": "Missing or empty title",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "The task",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Replaces title and/or done for the given task.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTask" },
              "example": { "title": "Buy oat milk", "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "400": {
            "description": "Invalid body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Task deleted, no content" },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Task stats",
        "description": "Returns total, done, and open task counts.",
        "responses": {
          "200": { "description": "Stats", "content": { "application/json": {} } }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset tasks",
        "description": "Restores the 3 example seed tasks. Useful for demos.",
        "responses": {
          "200": { "description": "Tasks reset", "content": { "application/json": {} } }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Buy milk" },
          "done": { "type": "boolean", "example": false }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "Buy milk" }
        }
      },
      "UpdateTask": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Buy oat milk" },
          "done": { "type": "boolean", "example": true }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Task 99 not found" }
        }
      }
    }
  }
}